requirements.txt
hyperlpr3==0.1.3
#pillow PIL
pillow==10.0.1
onnxruntime==1.8.1

# opencv-python==4.8.0.76
# numpy==1.24.4
# python-multipart==0.0.6
# uvicorn==0.23.2
# requests==2.31.0

PyQt5==5.15.9
#-------------

opencv-python==4.8.1.78
numpy~=1.24.4
matplotlib~=3.7.3
paddleocr~=2.7.0.3